requirements.txt
pynput
pystray
pillow